Split library.collectors into library.collectors.controller & library.collectors.others
