+++ metrics_utility/library/collectors/__init__.py
@@ -1,24 +1,7 @@
-from .config import config
-from .execution_environments import execution_environments
-from .job_host_summary import job_host_summary
-from .job_host_summary_service import job_host_summary_service
-from .main_host import main_host
-from .main_indirectmanagednodeaudit import main_indirectmanagednodeaudit
-from .main_jobevent import main_jobevent
-from .main_jobevent_service import main_jobevent_service
-from .total_workers_vcpu import total_workers_vcpu
-from .unified_jobs import unified_jobs
+from . import controller, others
 
 
 __all__ = [
-    'config',
-    'execution_environments',
-    'job_host_summary',
-    'job_host_summary_service',
-    'main_host',
-    'main_indirectmanagednodeaudit',
-    'main_jobevent',
-    'main_jobevent_service',
-    'total_workers_vcpu',
-    'unified_jobs',
+    'controller',
+    'others',
 ]
