+++ metrics_utility/library/collectors/controller/job_host_summary_service.py
@@ -1,4 +1,4 @@
-from .util import collector, copy_table
+from ..util import collector, copy_table
 
 
 @collector
