+++ metrics_utility/library/collectors/others/__init__.py
@@ -1,0 +1,6 @@
+from .total_workers_vcpu import total_workers_vcpu
+
+
+__all__ = [
+    'total_workers_vcpu',
+]
