+++ metrics_utility/library/collectors/others/total_workers_vcpu.py
@@ -5,8 +5,8 @@ from typing import Tuple
 
 from metrics_utility.logger import logger
 
+from ..util import collector
 from .prometheus_client import PrometheusClient
-from .util import collector
 
 
 @collector
