+++ metrics_utility/test/library/test_collectors_job_host_summary.py
@@ -2,7 +2,7 @@ import datetime
 
 from unittest.mock import MagicMock, patch
 
-from metrics_utility.library.collectors.job_host_summary import job_host_summary
+from metrics_utility.library.collectors.controller.job_host_summary import job_host_summary
 
 
 def test_job_host_summary_basic():
@@ -32,7 +32,7 @@ def test_job_host_summary_with_output_dir():
     assert instance.kwargs['output_dir'] == output_dir
 
 
-@patch('metrics_utility.library.collectors.job_host_summary.copy_table')
+@patch('metrics_utility.library.collectors.controller.job_host_summary.copy_table')
 def test_job_host_summary_calls_copy_table(mock_copy_table):
     """Test that job_host_summary calls copy_table with correct parameters."""
     mock_db = MagicMock()
@@ -53,7 +53,7 @@ def test_job_host_summary_calls_copy_table(mock_copy_table):
     assert result == ['/tmp/main_jobhostsummary_table.csv']
 
 
-@patch('metrics_utility.library.collectors.job_host_summary.copy_table')
+@patch('metrics_utility.library.collectors.controller.job_host_summary.copy_table')
 def test_job_host_summary_query_contains_time_range(mock_copy_table):
     """Test that the query includes the time range."""
     mock_db = MagicMock()
@@ -74,7 +74,7 @@ def test_job_host_summary_query_contains_time_range(mock_copy_table):
     assert 'main_jobhostsummary.modified <' in query
 
 
-@patch('metrics_utility.library.collectors.job_host_summary.copy_table')
+@patch('metrics_utility.library.collectors.controller.job_host_summary.copy_table')
 def test_job_host_summary_query_structure(mock_copy_table):
     """Test that the SQL query has expected structure."""
     mock_db = MagicMock()
@@ -98,7 +98,7 @@ def test_job_host_summary_query_structure(mock_copy_table):
     assert 'main_organization' in query
 
 
-@patch('metrics_utility.library.collectors.job_host_summary.copy_table')
+@patch('metrics_utility.library.collectors.controller.job_host_summary.copy_table')
 def test_job_host_summary_isoformat(mock_copy_table):
     """Test that datetime objects are converted to isoformat in query."""
     mock_db = MagicMock()
